Use POSIX::Spawn instead of Open3.capture3 for faster diffing
This is a much faster way to shell out in Ruby for reasons better
explained by other folks [1]. (It also has other features I plan to
make use of [2].)

The only part I'm not confident in is the encoding conversion. I'm not
sure why this is necessary under POSIX::Spawn but wasn't under Open3.
Maybe Open3 respects `Encoding.default_external` and POSIX::Spawn
doesn't?

In any case force-encoding the diff output seems to resolve the errors
I was getting. If encoding fails for any reason, we can let the caller
deal with any fallout.

Unfortunately I did have to remove a test from the test suite. This was
added in 2012 to handle a behavior in Ruby 1.9, so hopefully this isn't
a problem 10+ years later?

Credit for a lot of the ideas here goes to ZaiusInc who also forked
diffy and made similar changes [3].

[1] See charts in the readme: https://github.com/rtomayko/posix-spawn
[2] https://www.johnnunemaker.com/limit-everything-timeouts-for-commands-in-ruby/
[3] https://github.com/ZaiusInc/diffy/commit/0065c4ac32043b8acb989f0a654baf14b3503198