Add some safety options to set limits on diff process execution
Applying principles from [1], this allows passing `diff_timeout` and
`diff_max_length` option, which will be given to POSIX::Spawn as its
`timeout` and `max` options.

This is useful if user input is being diffed, and you want to protect
against a runaway diff command tying up e.g. web server resources.

Note that the max length is for output from the diff binary itself, not
the to-stringed or to-HTML'd output.

These options are turned off by default to maintain backwards
compatibility.

[1] https://www.johnnunemaker.com/limit-everything-timeouts-for-commands-in-ruby/